+++ src/components/feedback/__tests__/feedback-sent-text.test.js
@@ -39,10 +39,7 @@ describe('Sent text feedback', () => {
       .props()
       .onChange('cool beans!');
     // simulate submit button click
-    feedback
-      .find('[data-test="dr-ui--feedback-page-submit"]')
-      .at(0)
-      .simulate('click');
+    feedback.find('#dr-ui--feedback-page-submit').at(0).simulate('click');
 
     expect(feedback.state()).toEqual({
       event: {