+++ src/components/lazy/index.js
@@ -1,0 +1,3 @@
+import main from './lazy';
+
+export default main;