+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Lazy
+
 ## 2.3.0
 
 - Update `GlWrapper` component to check for Internet Explorer and provide reason for when the issue is not Internet Explorer or WebGL support. [#387](https://github.com/mapbox/dr-ui/pull/387)