+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Performance
 
-- Add Lazy component. [#392](https://github.com/mapbox/dr-ui/pull/392)
+- 🚨 Rename `GLWrapper` as `MapWrapper`.
+  - You must define the `height` of the map. This will set the height of the loader to prevent a content layout shift after the map loads.
+- `DemoIframe` has a required `title` prop to provide a description of the iframe's content.
 
 ## 2.3.0